Added MinedBlock, MiningJob models, set Transaction properties not to be serialized if null

+++ Node.Api/Models/Balance.cs
@@ -7,6 +7,6 @@ namespace Node.Api.Models
         public int Confirmations { get; set; }
 
         [Display(Name ="Balance")]
-        public int BalanceValue { get; set; }
+        public long BalanceValue { get; set; }
     }
 }
